refactor: split smoke hooks and consolidate smoke tooling
