+++ tests/smoke/run_mapgen_sweep_mac.sh
@@ -5,6 +5,8 @@ SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 RUNNER="$SCRIPT_DIR/run_lan_helo_chunk_smoke_mac.sh"
 HEATMAP="$SCRIPT_DIR/generate_mapgen_heatmap.py"
 AGGREGATE="$SCRIPT_DIR/generate_smoke_aggregate_report.py"
+COMMON_SH="$SCRIPT_DIR/lib/common.sh"
+source "$COMMON_SH"
 
 APP="$HOME/Library/Application Support/Steam/steamapps/common/Barony/Barony.app/Contents/MacOS/Barony"
 DATADIR=""
@@ -64,33 +66,21 @@ USAGE
 }
 
 is_uint() {
-	[[ "$1" =~ ^[0-9]+$ ]]
+	smoke_is_uint "$1"
 }
 
 log() {
-	printf '[%s] %s\n' "$(date '+%H:%M:%S')" "$*"
+	smoke_log "$*"
 }
 
 read_summary_key() {
 	local key="$1"
 	local file="$2"
-	local line
-	line="$(rg -n "^${key}=" "$file" | head -n 1 || true)"
-	if [[ -z "$line" ]]; then
-		echo ""
-		return
-	fi
-	echo "${line#*=}"
+	smoke_summary_get "$key" "$file"
 }
 
 count_fixed_lines() {
-	local file="$1"
-	local needle="$2"
-	if [[ ! -f "$file" ]]; then
-		echo 0
-		return
-	fi
-	rg -F -c "$needle" "$file" 2>/dev/null || echo 0
+	smoke_count_fixed_lines "$1" "$2"
 }
 
 while (($# > 0)); do
