+++ tests/smoke/run_save_reload_compat_smoke_mac.sh
@@ -10,6 +10,8 @@ OUTDIR=""
 
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 HELO_RUNNER="$SCRIPT_DIR/run_lan_helo_chunk_smoke_mac.sh"
+COMMON_SH="$SCRIPT_DIR/lib/common.sh"
+source "$COMMON_SH"
 
 usage() {
 	cat <<'USAGE'
@@ -27,21 +29,17 @@ USAGE
 }
 
 is_uint() {
-	[[ "$1" =~ ^[0-9]+$ ]]
+	smoke_is_uint "$1"
 }
 
 log() {
-	printf '[%s] %s\n' "$(date '+%H:%M:%S')" "$*"
+	smoke_log "$*"
 }
 
 read_summary_value() {
 	local summary_file="$1"
 	local key="$2"
-	if [[ ! -f "$summary_file" ]]; then
-		echo ""
-		return
-	fi
-	sed -n "s/^${key}=//p" "$summary_file" | tail -n 1
+	smoke_summary_get_last "$key" "$summary_file"
 }
 
 while (($# > 0)); do
@@ -204,7 +202,7 @@ fi
 	echo "SWEEP_LINE=$sweep_line"
 } > "$OUTDIR/summary.env"
 
-find "$LANE_OUTDIR/instances" -type f -name models.cache -delete 2>/dev/null || true
+smoke_prune_models_cache "$LANE_OUTDIR"
 
 log "summary=$OUTDIR/summary.env"
 log "csv=$CSV_PATH"
