+++ tests/smoke/run_splitscreen_cap_smoke_mac.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+COMMON_SH="$SCRIPT_DIR/lib/common.sh"
+source "$COMMON_SH"
+
 APP="$HOME/Library/Application Support/Steam/steamapps/common/Barony/Barony.app/Contents/MacOS/Barony"
 DATADIR=""
 WINDOW_SIZE="1280x720"
@@ -33,21 +37,15 @@ USAGE
 }
 
 is_uint() {
-	[[ "$1" =~ ^[0-9]+$ ]]
+	smoke_is_uint "$1"
 }
 
 log() {
-	printf '[%s] %s\n' "$(date '+%H:%M:%S')" "$*"
+	smoke_log "$*"
 }
 
 count_fixed_lines() {
-	local file="$1"
-	local needle="$2"
-	if [[ ! -f "$file" ]]; then
-		echo 0
-		return
-	fi
-	rg -F -c "$needle" "$file" 2>/dev/null || echo 0
+	smoke_count_fixed_lines "$1" "$2"
 }
 
 seed_smoke_home_profile() {
@@ -117,8 +115,7 @@ extract_latest_cap_status() {
 }
 
 prune_models_cache() {
-	local root="$1"
-	find "$root" -type f -name models.cache -delete 2>/dev/null || true
+	smoke_prune_models_cache "$1"
 }
 
 while (($# > 0)); do
